# AnimationCollectionAPI.DAL/DBO/Animation.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace AnimationCollectionAPI.DAL.DBO
{
    public class Animation : IEntity
    {
        public int Id { get; set; }
        [Required]
        public string Title { get; set; }
        public string Author { get; set; }
        public string Description { get; set; }
        [Required]
        public string Link { get; set; }
    }
}